requirements.txt
selenium>=4.0.0
webdriver-manager>=4.0.0
Pillow>=9.0.0
tqdm>=4.0.0
